refactor: auth connection
