+++ src/app/config/settings.py
@@ -1,7 +1,10 @@
 # infra/settings.py
 from pydantic_settings import BaseSettings
+from pydantic import SecretStr
+
 
 class Settings(BaseSettings):
+    database_url: SecretStr
     redis_host: str
     redis_port: int
     session_ttl: int
