+++ src/app/user/dto.py
@@ -1,4 +1,8 @@
 from sqlalchemy.orm import sessionmaker
+from pydantic import BaseModel
+from pydantic import SecretStr
+
+
 
 class User(BaseModel):
     username: str
